create a separate function module for initialization of database routes.
